@@ -37,46 +37,42 @@ dependencies:
   cupertino_icons: ^1.0.8 #ios 风格图标
   remixicon: ^1.2.0 # remix 图标
   get: ^4.6.6 #状态以及路由管理
-  dio: ^5.6.0 #HTTP请求
+  dio: ^5.7.0 #HTTP请求
   hive: ^2.2.3 #使用 AES-256 进行严格加密
   hive_flutter: ^1.1.0 #使用 AES-256 进行严格加密扩展功能
   crypton: ^2.2.1 #用于非对称加密和数字签名
-  crypto: ^3.0.5 #SHA、MD5 和 HMAC 加密函数的实现
+  crypto: ^3.0.6 #SHA、MD5 和 HMAC 加密函数的实现
   logger: ^2.4.0 #日志
   protobuf: ^3.1.0 #用于 protobuf协议
   intl: ^0.19.0 #多语言
   universal_html: ^2.2.4 #简化跨平台开发和 HTML/XML 处理
   html_unescape: ^2.0.0 #一个用于取消转义 HTML 的小型库
-  flutter_smart_dialog: ^4.9.8 #一个优雅的Flutter Dialog解决方案，轻松实现Toast、Loading和自定义Dialog
+  flutter_smart_dialog: ^4.9.8+3 #一个优雅的Flutter Dialog解决方案，轻松实现Toast、Loading和自定义Dialog
   flutter_staggered_grid_view: ^0.7.0 #提供了 Flutter 网格布局的集合
-  tab_indicator_styler: ^2.0.0 #直接将漂亮且流行的选项卡指示器添加到默认的 Flutter TabBar
-  multi_split_view: ^3.1.0 #为 Flutter 提供水平或垂直的多个分割视图。
-  extended_image: ^8.2.1 #官方扩展图像
+  multi_split_view: ^3.6.0 #为 Flutter 提供水平或垂直的多个分割视图。
+  extended_image: ^9.0.7 #官方扩展图像
   flutter_swiper_view:
     git: 
       url: https://github.com/whevether/flutter_swiper_view.git
       ref: main
   easy_refresh: ^3.4.0 #上拉下拉刷新
-  lottie: ^3.1.2 #lottie动画
+  lottie: ^3.1.3 #lottie动画
   photo_view: ^0.15.0 #照片视图提供了手势敏感的可缩放小部件
   preload_page_view: ^0.2.0 #预先加载page view
   scrollable_positioned_list: ^0.3.8 #带有辅助方法的列表，用于以编程方式滚动到某个项目
-  url_launcher: ^6.3.0 #打开链接
-  webview_flutter: ^4.9.0 #webview
-  package_info_plus: ^8.0.2 #包信息
-  flutter_widget_from_html_core: ^0.15.1 #html 转widget
-  share_plus: ^10.0.2 #分享
+  url_launcher: ^6.3.1 #打开链接
+  webview_flutter: ^4.10.0 #webview
+  package_info_plus: ^8.1.0 #包信息
+  flutter_widget_from_html_core: ^0.15.2 #html 转widget
+  share_plus: ^10.1.1 #分享
   permission_handler: ^11.3.1 #权限检查
-  path_provider: ^2.1.4 #常用目录
+  path_provider: ^2.1.5 #常用目录
   gal: ^2.3.0 #保存图片/视频
-  connectivity_plus: ^6.0.5 #连接状态
-  battery_plus: ^6.0.1 #电池状态
+  connectivity_plus: ^6.1.0 #连接状态
+  battery_plus: ^6.2.0 #电池状态
   file_selector: ^1.0.3 #文件选择
   windows_single_instance: ^1.0.1 #windows 单实例
-  intl_translation: #多语言生成
-    git: 
-      url: https://github.com/whevether/intl_translation.git
-      ref: main
+  intl_translation: ^0.20.0 #多语言生成
 dev_dependencies:
   flutter_test:
     sdk: flutter
