@@ -9,7 +9,7 @@ import 'package:wanime/widgets/status/app_empty_widget.dart';
 import 'package:get/get.dart';
 
 class ComicDownloadedView extends StatelessWidget {
-  const ComicDownloadedView({Key? key}) : super(key: key);
+  const ComicDownloadedView({super.key});
 
   @override
   Widget build(BuildContext context) {
