@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:wanime/routes/app_navigator.dart';
 
 class TestSubRoutePage extends StatelessWidget {
-  const TestSubRoutePage({Key? key}) : super(key: key);
+  const TestSubRoutePage({super.key});
 
   @override
   Widget build(BuildContext context) {
