@@ -9,7 +9,7 @@ import 'package:get/get.dart';
 import 'package:remixicon/remixicon.dart';
 
 class IndexPage extends GetView<IndexController> {
-  const IndexPage({Key? key}) : super(key: key);
+  const IndexPage({super.key});
 
   @override
   Widget build(BuildContext context) {
