@@ -7,7 +7,7 @@ import 'package:wanime/widgets/tab_appbar.dart';
 import 'package:get/get.dart';
 
 class NewsHomePage extends GetView<NewsHomeController> {
-  const NewsHomePage({Key? key}) : super(key: key);
+  const NewsHomePage({super.key});
 
   @override
   Widget build(BuildContext context) {
