@@ -121,8 +121,8 @@ class NovelDownloader {
       await _saveInfo();
     } catch (e) {
       Log.logPrint(e);
-      if (e is DioError) {
-        if (e.type == DioErrorType.cancel) rethrow;
+      if (e is DioException) {
+        if (e.type == DioExceptionType.cancel) rethrow;
         if (status == DownloadStatus.waitNetwork ||
             status == DownloadStatus.pauseCellular) rethrow;
         if (retryTime < 3) {
@@ -166,8 +166,8 @@ class NovelDownloader {
       await _saveInfo();
     } catch (e) {
       Log.logPrint(e);
-      if (e is DioError) {
-        if (e.type == DioErrorType.cancel) rethrow;
+      if (e is DioException) {
+        if (e.type == DioExceptionType.cancel) rethrow;
         if (status == DownloadStatus.waitNetwork ||
             status == DownloadStatus.pauseCellular) rethrow;
         if (retryTime < 3) {
