@@ -3,7 +3,7 @@ import 'package:wanime/app/app_style.dart';
 import 'package:lottie/lottie.dart';
 
 class AppLoaddingWidget extends StatelessWidget {
-  const AppLoaddingWidget({Key? key}) : super(key: key);
+  const AppLoaddingWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
