@@ -45,21 +45,21 @@ class MaterialHeader2 extends Header {
 
   const MaterialHeader2({
     this.key,
-    double triggerOffset = 100,
-    bool clamping = true,
-    IndicatorPosition position = IndicatorPosition.above,
-    Duration processedDuration = const Duration(milliseconds: 200),
-    physics.SpringDescription? spring,
-    bool springRebound = false,
+    super.triggerOffset = 100,
+    super.clamping = true,
+    super.position,
+    super.processedDuration = const Duration(milliseconds: 200),
+    super.spring,
+    super.springRebound = false,
     SpringBuilder? readySpringBuilder,
     FrictionFactor? frictionFactor,
-    bool safeArea = true,
-    double? infiniteOffset,
-    bool? hitOver,
-    bool? infiniteHitOver,
-    bool hapticFeedback = false,
-    bool triggerWhenRelease = false,
-    double maxOverOffset = double.infinity,
+    super.safeArea,
+    super.infiniteOffset,
+    super.hitOver,
+    super.infiniteHitOver,
+    super.hapticFeedback,
+    super.triggerWhenRelease,
+    super.maxOverOffset,
     required this.child,
     this.backgroundColor,
     this.color,
@@ -72,15 +72,10 @@ class MaterialHeader2 extends Header {
     this.bezierBackgroundAnimation = false,
     this.bezierBackgroundBounce = false,
   }) : super(
-          triggerOffset: triggerOffset,
-          clamping: clamping,
-          processedDuration: processedDuration,
-          spring: spring,
           readySpringBuilder: readySpringBuilder ??
               (bezierBackgroundAnimation
                   ? kBezierSpringBuilder
                   : kMaterialSpringBuilder),
-          springRebound: springRebound,
           frictionFactor: frictionFactor ??
               (showBezierBackground
                   ? kBezierFrictionFactor
@@ -89,14 +84,6 @@ class MaterialHeader2 extends Header {
               (showBezierBackground
                   ? kBezierHorizontalFrictionFactor
                   : kMaterialHorizontalFrictionFactor),
-          safeArea: safeArea,
-          infiniteOffset: infiniteOffset,
-          hitOver: hitOver,
-          infiniteHitOver: infiniteHitOver,
-          position: position,
-          hapticFeedback: hapticFeedback,
-          triggerWhenRelease: triggerWhenRelease,
-          maxOverOffset: maxOverOffset,
         );
 
   @override
@@ -159,21 +146,21 @@ class MaterialFooter2 extends Footer {
   final Widget child;
   const MaterialFooter2({
     this.key,
-    double triggerOffset = 100,
-    bool clamping = true,
-    IndicatorPosition position = IndicatorPosition.above,
-    Duration processedDuration = const Duration(milliseconds: 200),
-    physics.SpringDescription? spring,
+    super.triggerOffset = 100,
+    super.clamping = true,
+    super.position,
+    super.processedDuration = const Duration(milliseconds: 200),
+    super.spring,
     SpringBuilder? readySpringBuilder,
-    bool springRebound = false,
+    super.springRebound = false,
     FrictionFactor? frictionFactor,
-    bool safeArea = true,
-    double? infiniteOffset,
-    bool? hitOver,
-    bool? infiniteHitOver,
-    bool hapticFeedback = false,
-    bool triggerWhenRelease = false,
-    double maxOverOffset = double.infinity,
+    super.safeArea,
+    super.infiniteOffset = null,
+    super.hitOver,
+    super.infiniteHitOver,
+    super.hapticFeedback,
+    super.triggerWhenRelease,
+    super.maxOverOffset,
     required this.child,
     this.backgroundColor,
     this.color,
@@ -186,15 +173,10 @@ class MaterialFooter2 extends Footer {
     this.bezierBackgroundAnimation = false,
     this.bezierBackgroundBounce = false,
   }) : super(
-          triggerOffset: triggerOffset,
-          clamping: clamping,
-          processedDuration: processedDuration,
-          spring: spring,
           readySpringBuilder: readySpringBuilder ??
               (bezierBackgroundAnimation
                   ? kBezierSpringBuilder
                   : kMaterialSpringBuilder),
-          springRebound: springRebound,
           frictionFactor: frictionFactor ??
               (showBezierBackground
                   ? kBezierFrictionFactor
@@ -203,14 +185,6 @@ class MaterialFooter2 extends Footer {
               (showBezierBackground
                   ? kBezierHorizontalFrictionFactor
                   : kMaterialHorizontalFrictionFactor),
-          safeArea: safeArea,
-          infiniteOffset: infiniteOffset,
-          hitOver: hitOver,
-          infiniteHitOver: infiniteHitOver,
-          position: position,
-          hapticFeedback: hapticFeedback,
-          triggerWhenRelease: triggerWhenRelease,
-          maxOverOffset: maxOverOffset,
         );
 
   @override
@@ -285,7 +259,7 @@ class _MaterialIndicator extends StatefulWidget {
   final Widget child;
 
   const _MaterialIndicator({
-    Key? key,
+    super.key,
     required this.state,
     required this.disappearDuration,
     required this.reverse,
@@ -300,7 +274,7 @@ class _MaterialIndicator extends StatefulWidget {
     this.bezierBackgroundColor,
     this.bezierBackgroundAnimation = false,
     this.bezierBackgroundBounce = false,
-  }) : super(key: key);
+  });
 
   @override
   State<_MaterialIndicator> createState() => _MaterialIndicatorState();
