@@ -4,7 +4,7 @@ import 'package:lottie/lottie.dart';
 
 class AppEmptyWidget extends StatelessWidget {
   final Function()? onRefresh;
-  const AppEmptyWidget({this.onRefresh, Key? key}) : super(key: key);
+  const AppEmptyWidget({this.onRefresh, super.key});
 
   @override
   Widget build(BuildContext context) {
